Requirements

+++ external/TrackEval/trackeval/datasets/mot_challenge_2d_box.py
@@ -298,7 +298,7 @@ class MotChallenge2DBox(_BaseDataset):
             time_key = str(t + 1)
             if time_key in read_data.keys():
                 try:
-                    time_data = np.asarray(read_data[time_key], dtype=np.float)
+                    time_data = np.asarray(read_data[time_key], dtype=np.float32)
                 except ValueError:
                     if is_gt:
                         raise TrackEvalException(
@@ -454,7 +454,7 @@ class MotChallenge2DBox(_BaseDataset):
 
             # Match tracker and gt dets (with hungarian algorithm) and remove tracker dets which match with gt dets
             # which are labeled as belonging to a distractor class.
-            to_remove_tracker = np.array([], np.int)
+            to_remove_tracker = np.array([], np.int32)
             if (
                 self.do_preproc
                 and self.benchmark != "MOT15"
@@ -528,7 +528,7 @@ class MotChallenge2DBox(_BaseDataset):
             gt_id_map[unique_gt_ids] = np.arange(len(unique_gt_ids))
             for t in range(raw_data["num_timesteps"]):
                 if len(data["gt_ids"][t]) > 0:
-                    data["gt_ids"][t] = gt_id_map[data["gt_ids"][t]].astype(np.int)
+                    data["gt_ids"][t] = gt_id_map[data["gt_ids"][t]].astype(np.int32)
         if len(unique_tracker_ids) > 0:
             unique_tracker_ids = np.unique(unique_tracker_ids)
             tracker_id_map = np.nan * np.ones((np.max(unique_tracker_ids) + 1))
@@ -537,7 +537,7 @@ class MotChallenge2DBox(_BaseDataset):
                 if len(data["tracker_ids"][t]) > 0:
                     data["tracker_ids"][t] = tracker_id_map[
                         data["tracker_ids"][t]
-                    ].astype(np.int)
+                    ].astype(np.int32)
 
         # Record overview statistics.
         data["num_tracker_dets"] = num_tracker_dets
